requirements.txt
django
psycopg2
celery
redis
djangorestframework
django-anymail[mailgun]
django-crispy-forms
django-autocomplete-light